schlong

+++ src/main/java/finalproject/alongtheway/AlongTheWayController.java
@@ -14,7 +14,7 @@ import finalproject.alongtheway.waypoints.Steps;
 
 @Controller
 public class AlongTheWayController {
-	
+
 	@Autowired
 	private GoogleApiService googleApiService;
 
@@ -31,7 +31,7 @@ public class AlongTheWayController {
 		ModelAndView mav = new ModelAndView("info");
 		return mav;
 	}
-	
+
 	@RequestMapping("/results")
 	public ModelAndView results(@RequestParam(name = "location", required = true) String location) {
 		List<Businesses> results;
@@ -53,25 +53,24 @@ public class AlongTheWayController {
 
 		return mav;
 	}
-	
+
 	@RequestMapping("/directions")
 	public ModelAndView direction() {
 		List<Steps> steps;
 		Steps step;
-		//yo, peep that 7, mang(that is the 8th step in the route, mang)
-		step = googleApiService.getWaypoints("","").get(7);
-		Double lat1 = step.getStartLocation().getStartLat();
-		Double lat2 = step.getEndLocation().getEndLat();
-		Double long1 = step.getStartLocation().getStartLong();
-		Double long2 = step.getEndLocation().getEndLong();
-		
-		
+		// yo, peep that 7, mang(that is the 8th step in the route, mang)
+		step = googleApiService.getWaypoints("", "").get(7);
+		Long lat1 = step.getStartLocation().getStartLat();
+		Long lat2 = step.getEndLocation().getEndLat();
+		Long long1 = step.getStartLocation().getStartLong();
+		Long long2 = step.getEndLocation().getEndLong();
+
 		ModelAndView mav = new ModelAndView("directions");
 		mav.addObject("lat1", lat1);
 		mav.addObject("lat2", lat2);
 		mav.addObject("long1", long1);
 		mav.addObject("long2", long2);
-	
+
 		return mav;
 	}
